feat!: add filtering to different commands

+++ install.sh
@@ -9,11 +9,21 @@ GREEN="\e[32;1m"
 
 destination="${1:-"${HOME}/.local/bin"}"
 
-# Install templates and backends
-./pm install-hook
+# Path were pm stores its templates and backends
+PM_DATA_DIR="${HOME}/.local/share/pm"
 
-if [[ ! -f "${destination}/pm" ]]; then
-    command cp -i pm "${destination}/pm"
-fi
+[[ ! -d "${PM_DATA_DIR}" ]] && command mkdir -p "${PM_DATA_DIR}"
+
+# Copy default templates and backends
+command cp -R backends "${PM_DATA_DIR}/"
+command cp -R templates "${PM_DATA_DIR}/"
+
+# Sets the default global template
+echo "default" > "${PM_DATA_DIR}/global-template"
+
+# Sets the default global backend
+echo "vscode" > "${PM_DATA_DIR}/global-backend"
+
+[[ ! -f "${destination}/pm" ]] && command cp -i pm "${destination}/pm"
 
 echo -e "${GREEN}✓${WHITE} pm is ready to use"
